pyproject.toml
[tool.poetry]
name = "evo-omni-publisher-engine"
version = "0.1.0"
description = ""
authors = ["Manuel Sandoval <msandovalgm@live.com.mx>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.129.2"
uvicorn = "^0.41.0"
sqlalchemy = "^2.0.46"
psycopg2-binary = "^2.9.11"
python-dotenv = "^1.2.1"
apscheduler = "^3.11.2"
boto3 = "^1.42.54"
google-api-python-client = "^2.190.0"
google-auth-httplib2 = "^0.3.0"
google-auth-oauthlib = "^1.2.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
